Refine binding of Vividus meta to Allure labels (#499)
Links in Allure report are generated from the following meta:
- @testCaseId
- @issueId
- @requirementId

The corresponding properties to define link URL patterns are:
- system.allure.link.tms.pattern
- system.allure.link.issue.pattern
- system.allure.link.requirement.pattern

Meta @testTier is deprecated in favor of new meta @severity to
keep consistency in the reporting. The acceptable range of values
is left unchanged: 1..5.

Meta @group is no more mapped to Allure labels since it's not
used in the reporting. But it's still can be used by users to
mark stories and scenarios for filtering.

Meta @epic and @feature are reserved for future use in Allure
Behaviors plugin.