[bdd-engine] Fix resolving of circular variable references (#711)
The previous fix #700 was not good-enough, since it wasn't able
to handle circular variable references in complex expressions
when the value to process had prefix/postfix before/after the
actual variable. Since there is no way to parse and detect
circular variables, this solution introduces max nested level of
variables (value is 16). If max level is reached, then result of
one resolution cycle is returned.