[plugin-mobile-app] Deprecate web view switching by index
The order of web views returned by appium is not reliable.
For android platform the web views are retrieved by parsing UNIX socket names from /proc/net/unix.
For ios platform the web views are consideres as separate processes within the active application.
Sorting the web view by Inode number for android or PID for ios doen't seem logical for end users and as practice shows the order is not guaranteed.