[tests] Disable idempotence for test Kafka producer
https://kafka.apache.org/30/documentation.html#producerconfigs:
"Notable changes in 3.0.1
Idempotence for the producer is enabled by default if no conflicting configurations are set. A bug prevented the producer idempotence default from being applied which meant that it remained disabled unless the user had explicitly set enable.idempotence to true. See KAFKA-13598for more details. This issue was fixed and the default is properly applied."

BUT enabled idempotence for producer causes "org.apache.kafka.common.errors.ClusterAuthorizationException: Cluster authorization failed." for CloudKarafka (https://stackoverflow.com/questions/67148019/kafka-producer-property-enable-idempotence-true-is-causing-error), so it's needed to be disabled.