[engine] Fix parsing of nested expressions without parameters (#3582)
The previous `RELUCTANT_EXPRESSION_PATTERN` didn't correctly process expressions with the number of characters which was not multiple of 3.
The previous regex correctly matched nested expressions without arguments if the number of characters in this expression is multiple of three only  (3+3+3...). For example, it correctly parsed "story_random" or "target", but not "story_indexed" or "targets". 

Co-authored-by: draker94 <noreply@github.com>