[tests] Use HTTPS instead of HTTP in proxy steps tests
New Headless is the real Chrome browser, and is thus more authentic. Chrome introduced support for the "Upgrade-Insecure-Requests" header in version 43, which was released in 2015. This header is part of an effort to improve security by requesting that servers upgrade insecure resources to secure ones when a webpage is loaded over HTTPS. See more: https://stackoverflow.com/a/31570584/2067574